Created javascript.js with rules on playing rps
The computer selects a choice with the 'computerPlay()' function. That
choice is then given to 'playRound()' funciton along side a starting
user choice.

The 'playRound()' funcion lets the user know if the computer or the
user won the round or if the game was a tie.

Things needed still to be added,

    - Ask user for choice

    - Keep track of score  (make winner first to 5)